feat(magento): onLoad refresh is optional
